Fix missing badge for Turbo Drive page loads
Git bisect pointed to turbo-rails repo commit d18b2ee being the cause
of this bug. Previously Turbo Drive requests were displayed via the callback
added to the `window.fetch`. That turbo-rails commit however change it
so that a reference to `window.fetch` was stored when Turbo JS is loaded (before
rack-mini-profiler has added the then callback) meaning that the
callback to display those profiling request results never executes.

Adding a callback to `window.Turbo.fetch` is difficult as `window.Turbo`
is a frozen object, so instead a turbo event listener is used.